A bunch of template hacking; index "works", course pages need more
The main idea is to use racket web-server/templates with backtick (`) as the
templating character. (@ is a pain because it is used for email addresses). All
main/index HTML files will be passed through the templater, which can fill in
the menubar (and other variables, if needed, in the future).

A bunch of images needed are not in this repo, so I've made them refer to
Bootstrapworld.org/images directly; they could be downloaded and copied into a
suitable location within this repo.